requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
avalanche-lib==0.4.0
beautifulsoup4==4.12.2
blessed==1.20.0
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.16.0
cmake==3.25.0
confection==0.1.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
datasets==2.14.6
dill==0.3.7
Django==4.2.6
docker-pycreds==0.4.0
evaluate==0.4.1
fairscale==0.4.13
ffmpy==0.3.1
filelock==3.9.0
fire==0.5.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.1
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.1.0
gpustat==1.1.1
GPUtil==1.4.0
gradio==3.47.1
gradio_client==0.6.0
graphviz==0.20.1
grpcio==1.59.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
image==1.5.33
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.1.0
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
lightning-utilities==0.9.0
lit==15.0.7
Markdown==3.5
MarkupSafe==2.1.2
matplotlib==3.8.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
networkx==3.0
ninja==1.11.1.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.108
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
orjson==3.9.7
pandas==2.1.1
pathtools==0.1.2
Pillow==9.3.0
preshed==3.0.9
protobuf==4.23.4
psutil==5.9.6
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydub==0.25.1
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytorchcv==0.0.67
pytz==2023.3.post1
PyYAML==6.0.1
quadprog==0.1.11
referencing==0.30.2
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rpds-py==0.10.4
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.32.0
setproctitle==1.3.3
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
starlette==0.27.0
sympy==1.12
tensorboard==2.15.0
tensorboard-data-server==0.7.1
tensorboardX==1.8
termcolor==2.3.0
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.5.4
tokenizers==0.14.1
toolz==0.12.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchdata==0.7.1
torchmetrics==1.2.0
torchvision==0.15.2+cu117
torchviz==0.0.2
tqdm==4.66.1
transformers @ git+https://github.com/huggingface/transformers@897a826d830e8b1e03eb482b165b5d88a7a08d5f
triton==2.0.0
typer==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
wandb==0.15.12
wasabi==1.1.2
weasel==0.3.3
websockets==11.0.3
Werkzeug==3.0.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
